requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==3.2.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
envparse==0.2.0
fastapi==0.115.11
filelock==3.18.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
Mako==1.3.9
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
platformdirs==4.3.8
pluggy==1.5.0
pre_commit==4.2.0
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.26.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.31.2
